+++ tests/testthat/test-partial_eval.R
@@ -18,3 +18,8 @@ test_that("namespace operators always evaluated locally", {
 test_that("namespaced calls to dplyr functions are stripped", {
   expect_equal(partial_eval(quote(dplyr::n())), expr(n()))
 })
+
+test_that("use quosure environment for unevaluted formulas", {
+  x <- 1
+  expect_equal(partial_eval(expr(~x)), ~1)
+})