+++ NEWS.md
@@ -1,12 +1,50 @@
 # dbplyr 1.2.1.9001
 
+* Fixes translation of `cummean()` from 'mean' to 'avg' (#157)
+
+* Generalizes `win_cummulative()` so that it works with MS SQL back ends 
+
+* Adds tests for `cummean()`, `cummax()` and `cummin()`
+
+* New translation for `bit64::as.integer64()` (#3305)
+
+* `case_when` now translates with a ELSE clause if a formula of the form `TRUE~<RHS>`
+is provided . (@cderv, #112)
+
+* SQLite gains correct translation for `as.numeric()`/`as.double()` 
+  (@chris-park, #171).
+
+* New translation for `bit64::as.integer64()` (#3305)
+
+* `case_when` now create a `ELSE` clause if a formula of the form `TRUE ~ <RHS>` 
+  is provided . (@cderv, #112)
+
+* `copy_to()` will only remove existing table when `overwrite = TRUE` and the
+  table already exists, eliminating a confusion "NOTICE" from PostgreSQL 
+  (#3197).
+
+* `vars` argument to `tbl_sql()` has been formally deprecated; it hasn't 
+  actually done anything for a while (#3254).
+
+* Adds custom `setdiff()` for Oracle connections (#3493)
+
+* `pull.tbl_sql()` now extracts correctly from grouped tables (#3562).
+
+* Custom `db_explain()` for Oracle connections (#3471)
+
 * `src` and `tbl` objects now include a class generated from the class of 
   the underlying connection object. This makes it possible for dplyr backends 
   to implement different behaviour at the dplyr level, when needed. (#2293)
 
 * Works on R 3.1
 
+* Fixes default parameter order for the `str_detect()` translation (#3397)
+
+* `sql_render.op()` now correctly forwards the `con` argument (@kevinykuo, #73).
+
 * Redshift `substr()` compatibility issue resolved (#3339)
+
+* `x %in% y` is now translated to `FALSE` if `y` is empty (@mgirlich, #160).
 
 # dbplyr 1.2.1
 
@@ -118,6 +156,8 @@
 *   SQLite: improved `na_if` translation (@cwarden)
 
 *   PostgreSQL: translation for `grepl()` added (@zozlak)
+
+*   Oracle: changed VARVHAR to VARCHAR2 datatype (@washcycle, #66)
 
 # dbplyr 1.1.0